// examples/onyx/types.ts
// AUTO-GENERATED BY onyx-gen. DO NOT EDIT.
// Generated at 2026-01-10T04:38:24.019Z

export interface User {
  id?: string;
  asdf?: string;
  username?: string;
  email?: string;
  isActive?: boolean;
  lastLoginAt: Date | null;
  createdAt?: Date;
  updatedAt?: Date;
  deletedAt: Date | null;
  [key: string]: any;
}

export interface UserProfile {
  id?: string;
  userId?: string;
  firstName?: string;
  lastName?: string;
  phone: string | null;
  address: any | null;
  avatarUrl: string | null;
  bio: string | null;
  createdAt?: Date;
  updatedAt: Date | null;
  deletedAt: Date | null;
  age: number | null;
  [key: string]: any;
}

export interface Role {
  id?: string;
  name?: string;
  description: string | null;
  isSystem?: boolean;
  createdAt?: Date;
  updatedAt?: Date;
  deletedAt: Date | null;
  [key: string]: any;
}

export interface Permission {
  id?: string;
  name?: string;
  description: string | null;
  createdAt?: Date;
  updatedAt?: Date;
  deletedAt: Date | null;
  [key: string]: any;
}

export interface UserRole {
  id?: string;
  userId?: string;
  roleId?: string;
  createdAt?: Date;
  [key: string]: any;
}

export interface RolePermission {
  id?: string;
  roleId?: string;
  permissionId?: string;
  createdAt?: Date;
  [key: string]: any;
}

export interface AuditLog {
  id?: string;
  dateTime?: Date;
  action: string | null;
  status: string | null;
  actorId: string | null;
  tenantId: string | null;
  targetId: string | null;
  resource: string | null;
  requestId: string | null;
  errorCode: string | null;
  errorMessage: string | null;
  changes: any | null;
  metadata: any | null;
  [key: string]: any;
}

export type OnyxSchema = {
  User: User;
  UserProfile: UserProfile;
  Role: Role;
  Permission: Permission;
  UserRole: UserRole;
  RolePermission: RolePermission;
  AuditLog: AuditLog;
};

export type Schema = OnyxSchema;
export const Schema = {} as OnyxSchema;

export enum tables {
  User = "User",
  UserProfile = "UserProfile",
  Role = "Role",
  Permission = "Permission",
  UserRole = "UserRole",
  RolePermission = "RolePermission",
  AuditLog = "AuditLog",
}